requirements.txt
chromadb
openai
python-telegram-bot
python-dotenv
pandas
tqdm
requests 